phase5_deployment/backend/requirements.txt
# Phase 5: Production Backend
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0

# Core (Phase 1-4)
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.2.0
statsmodels>=0.14.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Monitoring
prometheus-client>=0.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
